+++ CHANGELOG.md
@@ -9,9 +9,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- Local time now shown along eve time in timer list
+
 ### Changed
 
 ### Fixed
+
+- Form validation no longer forgets values for structure and solar system [#1](https://gitlab.com/ErikKalkoken/aa-structuretimers/-/issues/1)
 
 ## 1.0.0b2 - 2020-08-19